role + receiver
1. "role"

"role" adds another dimension to owns, burned, and balance.

previously these relationships used to be for "sender" only. But now it can handle "receiver" condition as well.

for example "the token will be minted only if the receiver has burned token X".

2. dynamic receiver

if a token does not specify a receiver, it can be gifted to anyone by passing the "receiver" attribute as the second parameter